Revert "Merge pull request #40 from hunner/chomp_lineendings"
This reverts commit 4607f82061e6748f23ef798ce62216f6da1f3ff7, reversing
changes made to 933366666d9875bfc20e050206f11616e795ece5.

Chomping line endings causes any tests that were already depending on
the line endings to fail. It is also conceptually wrong for a testing
framework to modify the output that is going to be used for asserting
against. Dealing with the line ending should be done in a custom
assertion or matcher.